+++ Solutions/Business.StarterIntranet/app/src/components/NotificationBanner/Notification.ts
@@ -8,10 +8,14 @@ class Notification implements INotification {
     public static SelectFields = [
         "Title",
         "IntranetNotificationDescription",
+        "IntranetNotificationBgColor",
+        "IntranetNotificationTextColor"
     ];
 
     public Title: string;
     public IntranetNotificationDescription: string;
+    public IntranetNotificationBgColor: string;
+    public IntranetNotificationTextColor: string;   
 }
 
 export default Notification;