+++ Solutions/Migration.Utilities.Discussions/Migration.Utilities.Dsicussions/ConnectionForm.Designer.cs
@@ -1,0 +1,231 @@
+﻿namespace SP_Discussion_Migrator
+{
+    partial class ConnectionForm
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            this.label1 = new System.Windows.Forms.Label();
+            this.siteURLTextbox = new System.Windows.Forms.TextBox();
+            this.usernameTextbox = new System.Windows.Forms.TextBox();
+            this.label2 = new System.Windows.Forms.Label();
+            this.label3 = new System.Windows.Forms.Label();
+            this.passwordTextbox = new System.Windows.Forms.TextBox();
+            this.connectButton = new MetroFramework.Controls.MetroButton();
+            this.exitButton = new MetroFramework.Controls.MetroButton();
+            this.radioButton1 = new System.Windows.Forms.RadioButton();
+            this.radioButton2 = new System.Windows.Forms.RadioButton();
+            this.label4 = new System.Windows.Forms.Label();
+            this.panel1 = new System.Windows.Forms.Panel();
+            this.panel1.SuspendLayout();
+            this.SuspendLayout();
+            // 
+            // label1
+            // 
+            this.label1.AutoSize = true;
+            this.label1.Location = new System.Drawing.Point(74, 108);
+            this.label1.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
+            this.label1.Name = "label1";
+            this.label1.Size = new System.Drawing.Size(50, 13);
+            this.label1.TabIndex = 0;
+            this.label1.Text = "Site URL";
+            this.label1.TextAlign = System.Drawing.ContentAlignment.MiddleRight;
+            // 
+            // siteURLTextbox
+            // 
+            this.siteURLTextbox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this.siteURLTextbox.Location = new System.Drawing.Point(127, 106);
+            this.siteURLTextbox.Margin = new System.Windows.Forms.Padding(2);
+            this.siteURLTextbox.Name = "siteURLTextbox";
+            this.siteURLTextbox.Size = new System.Drawing.Size(253, 20);
+            this.siteURLTextbox.TabIndex = 0;
+            // 
+            // usernameTextbox
+            // 
+            this.usernameTextbox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this.usernameTextbox.Location = new System.Drawing.Point(127, 127);
+            this.usernameTextbox.Margin = new System.Windows.Forms.Padding(2);
+            this.usernameTextbox.Name = "usernameTextbox";
+            this.usernameTextbox.Size = new System.Drawing.Size(253, 20);
+            this.usernameTextbox.TabIndex = 1;
+            // 
+            // label2
+            // 
+            this.label2.AutoSize = true;
+            this.label2.Location = new System.Drawing.Point(68, 129);
+            this.label2.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
+            this.label2.Name = "label2";
+            this.label2.Size = new System.Drawing.Size(55, 13);
+            this.label2.TabIndex = 2;
+            this.label2.Text = "Username";
+            this.label2.TextAlign = System.Drawing.ContentAlignment.MiddleRight;
+            // 
+            // label3
+            // 
+            this.label3.AutoSize = true;
+            this.label3.Location = new System.Drawing.Point(71, 149);
+            this.label3.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
+            this.label3.Name = "label3";
+            this.label3.Size = new System.Drawing.Size(53, 13);
+            this.label3.TabIndex = 2;
+            this.label3.Text = "Password";
+            this.label3.TextAlign = System.Drawing.ContentAlignment.MiddleRight;
+            // 
+            // passwordTextbox
+            // 
+            this.passwordTextbox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            | System.Windows.Forms.AnchorStyles.Right)));
+            this.passwordTextbox.Location = new System.Drawing.Point(127, 148);
+            this.passwordTextbox.Margin = new System.Windows.Forms.Padding(2);
+            this.passwordTextbox.Name = "passwordTextbox";
+            this.passwordTextbox.Size = new System.Drawing.Size(253, 20);
+            this.passwordTextbox.TabIndex = 2;
+            this.passwordTextbox.UseSystemPasswordChar = true;
+            // 
+            // connectButton
+            // 
+            this.connectButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
+            this.connectButton.Highlight = true;
+            this.connectButton.Location = new System.Drawing.Point(241, 220);
+            this.connectButton.Margin = new System.Windows.Forms.Padding(2);
+            this.connectButton.Name = "connectButton";
+            this.connectButton.Size = new System.Drawing.Size(85, 21);
+            this.connectButton.TabIndex = 3;
+            this.connectButton.Text = "C&onnect";
+            this.connectButton.UseSelectable = true;
+            this.connectButton.UseVisualStyleBackColor = true;
+            this.connectButton.Click += new System.EventHandler(this.connectButton_Click);
+            // 
+            // exitButton
+            // 
+            this.exitButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
+            this.exitButton.DialogResult = System.Windows.Forms.DialogResult.Cancel;
+            this.exitButton.Location = new System.Drawing.Point(330, 220);
+            this.exitButton.Margin = new System.Windows.Forms.Padding(2);
+            this.exitButton.Name = "exitButton";
+            this.exitButton.Size = new System.Drawing.Size(85, 21);
+            this.exitButton.TabIndex = 4;
+            this.exitButton.Text = "E&xit";
+            this.exitButton.UseSelectable = true;
+            this.exitButton.UseVisualStyleBackColor = true;
+            this.exitButton.Click += new System.EventHandler(this.exitButton_Click);
+            // 
+            // radioButton1
+            // 
+            this.radioButton1.AutoSize = true;
+            this.radioButton1.Checked = true;
+            this.radioButton1.Location = new System.Drawing.Point(5, 1);
+            this.radioButton1.Margin = new System.Windows.Forms.Padding(1);
+            this.radioButton1.Name = "radioButton1";
+            this.radioButton1.Size = new System.Drawing.Size(59, 17);
+            this.radioButton1.TabIndex = 5;
+            this.radioButton1.TabStop = true;
+            this.radioButton1.Text = "Source";
+            this.radioButton1.UseVisualStyleBackColor = true;
+            // 
+            // radioButton2
+            // 
+            this.radioButton2.AutoSize = true;
+            this.radioButton2.Location = new System.Drawing.Point(5, 19);
+            this.radioButton2.Margin = new System.Windows.Forms.Padding(1);
+            this.radioButton2.Name = "radioButton2";
+            this.radioButton2.Size = new System.Drawing.Size(56, 17);
+            this.radioButton2.TabIndex = 6;
+            this.radioButton2.Text = "Target";
+            this.radioButton2.UseVisualStyleBackColor = true;
+            // 
+            // label4
+            // 
+            this.label4.AutoSize = true;
+            this.label4.Location = new System.Drawing.Point(39, 65);
+            this.label4.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
+            this.label4.Name = "label4";
+            this.label4.Size = new System.Drawing.Size(88, 13);
+            this.label4.TabIndex = 0;
+            this.label4.Text = "Connection Type";
+            this.label4.TextAlign = System.Drawing.ContentAlignment.MiddleRight;
+            // 
+            // panel1
+            // 
+            this.panel1.Controls.Add(this.radioButton1);
+            this.panel1.Controls.Add(this.radioButton2);
+            this.panel1.Location = new System.Drawing.Point(127, 63);
+            this.panel1.Name = "panel1";
+            this.panel1.Size = new System.Drawing.Size(250, 38);
+            this.panel1.TabIndex = 5;
+            // 
+            // ConnectionForm
+            // 
+            this.AcceptButton = this.connectButton;
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.CancelButton = this.exitButton;
+            this.ClientSize = new System.Drawing.Size(430, 260);
+            this.Controls.Add(this.panel1);
+            this.Controls.Add(this.exitButton);
+            this.Controls.Add(this.connectButton);
+            this.Controls.Add(this.passwordTextbox);
+            this.Controls.Add(this.label3);
+            this.Controls.Add(this.usernameTextbox);
+            this.Controls.Add(this.label2);
+            this.Controls.Add(this.siteURLTextbox);
+            this.Controls.Add(this.label4);
+            this.Controls.Add(this.label1);
+            this.Margin = new System.Windows.Forms.Padding(2);
+            this.MaximizeBox = false;
+            this.MinimumSize = new System.Drawing.Size(430, 260);
+            this.Name = "ConnectionForm";
+            this.Padding = new System.Windows.Forms.Padding(13, 60, 13, 13);
+            this.ShowIcon = false;
+            this.Text = "Specify Connection details...";
+            this.Theme = MetroFramework.MetroThemeStyle.Default;
+            this.Load += new System.EventHandler(this.ConnectionForm_Load);
+            this.panel1.ResumeLayout(false);
+            this.panel1.PerformLayout();
+            this.ResumeLayout(false);
+            this.PerformLayout();
+
+        }
+
+        #endregion
+
+        private System.Windows.Forms.Label label1;
+        private System.Windows.Forms.TextBox siteURLTextbox;
+        private System.Windows.Forms.TextBox usernameTextbox;
+        private System.Windows.Forms.Label label2;
+        private System.Windows.Forms.Label label3;
+        private System.Windows.Forms.TextBox passwordTextbox;
+        private MetroFramework.Controls.MetroButton connectButton;
+        private MetroFramework.Controls.MetroButton exitButton;
+        private System.Windows.Forms.RadioButton radioButton1;
+        private System.Windows.Forms.RadioButton radioButton2;
+        private System.Windows.Forms.Label label4;
+        private System.Windows.Forms.Panel panel1;
+    }
+}
+